Update dev-env playbook
There are a few checks done on the existence of attributes of the
participants variable during the creation of users.
If a user is not registered in github, they are skipped.

If a user does not have their keys registered, they are not skipped
but no check is currently done for this.